requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
openai==1.3.0
python-multipart==0.0.6
anthropic==0.8.1
httpx==0.25.2 